refactor(admin): split admin into module
